feature: move (initial) serving of cached route to event handler
