+++ src/runtime/server/hooks/request.ts
@@ -1,24 +1,14 @@
 import { type H3Event } from 'h3'
-import type { NuxtMultiCacheSSRContext, RouteCacheItem } from '../../types'
+import type { NuxtMultiCacheSSRContext } from '../../types'
 import {
   MULTI_CACHE_CDN_CONTEXT_KEY,
   MULTI_CACHE_CONTEXT_KEY,
   MULTI_CACHE_PREFIX_KEY,
   MULTI_CACHE_ROUTE_CONTEXT_KEY,
-  encodeRouteCacheKey,
-  getCacheKeyWithPrefix,
 } from '../../helpers/server'
 import { NuxtMultiCacheRouteCacheHelper } from '../../helpers/RouteCacheHelper'
-import {
-  decodeRouteCacheItem,
-  handleRawCacheData,
-} from '../../helpers/cacheItem'
-import { logger } from '../../helpers/logger'
 import { useMultiCacheApp } from '../utils/useMultiCacheApp'
 import { NuxtMultiCacheCDNHelper } from '../../helpers/CDNHelper'
-import { serveCachedRoute } from '../../helpers/routeCache'
-import type { MultiCacheState } from '../../helpers/MultiCacheState'
-import { useRuntimeConfig } from '#imports'
 
 /**
  * Add the cache context singleton to the current request.
@@ -92,29 +82,6 @@ function applies(path: string): boolean {
   return !/.\.(ico|png|jpg|js|css|html|woff|woff2|ttf|otf|eot|svg)$/.test(path)
 }
 
-function canBeServedFromCache(
-  key: string,
-  decoded: RouteCacheItem,
-  state: MultiCacheState,
-): boolean {
-  const now = Date.now() / 1000
-  const isExpired = decoded.expires ? now >= decoded.expires : false
-
-  // Item is not expired, so we can serve it.
-  if (!isExpired) {
-    return true
-  }
-
-  // The route may be served stale while revalidating and it is not currently
-  // being revalidated.
-  if (decoded.staleWhileRevalidate && state.isBeingRevalidated(key)) {
-    return true
-  }
-
-  // Is both expired and not eligible to be served stale while revalidating.
-  return false
-}
-
 /**
  * Callback for the 'request' nitro hook.
  *
@@ -140,70 +107,5 @@ export async function onRequest(event: H3Event) {
   }
 
   // Add the cache context.
-  const multiCache = await addCacheContext(event)
-
-  // Route caching is not enabled, so we can return now.
-  if (!multiCache?.route) {
-    return
-  }
-
-  try {
-    const { serverOptions, state } = useMultiCacheApp()
-
-    // Build the cache key.
-    const fullKey = serverOptions?.route?.buildCacheKey
-      ? serverOptions.route.buildCacheKey(event)
-      : getCacheKeyWithPrefix(encodeRouteCacheKey(event.path), event)
-
-    // Check if there is a cache entry for this key.
-    const cachedRaw = handleRawCacheData(
-      await multiCache.route.getItemRaw(fullKey),
-    )
-
-    // No cache entry.
-    if (!cachedRaw) {
-      return
-    }
-
-    const decoded = decodeRouteCacheItem(cachedRaw)
-
-    // Decoding failed. May happen if the format is wrong, possibly after a
-    // deployment with a newer version.
-    if (!decoded) {
-      return
-    }
-
-    // Check if item can be served from cache.
-    if (!canBeServedFromCache(fullKey, decoded, state)) {
-      // Mark the key as being revalidated.
-      if (decoded.staleWhileRevalidate) {
-        state.addKeyBeingRevalidated(fullKey)
-        event.__MULTI_CACHE_REVALIDATION_KEY = fullKey
-      }
-
-      if (decoded.staleIfErrorExpires) {
-        // Store the decoded cache item in the event context.
-        // May be used by the error hook handler to serve a stale route on error.
-        event.__MULTI_CACHE_DECODED_CACHED_ROUTE = decoded
-      }
-
-      // Returning, so the route is revalidated.
-      return
-    }
-
-    const debugEnabled = useRuntimeConfig().multiCache.debug
-
-    if (debugEnabled) {
-      logger.info('Serving cached route for path: ' + event.path, {
-        fullKey,
-      })
-    }
-
-    await serveCachedRoute(event, decoded)
-  } catch (e) {
-    if (e instanceof Error) {
-      // eslint-disable-next-line no-console
-      console.debug(e.message)
-    }
-  }
+  await addCacheContext(event)
 }
